Move server app's s3 config file from utilities directory to config directory
