+++ app/server/app/config/s3.js
